Refactoring: extracted new helper class to handle matching of expectations on equivalent/replacement instances.
