+++ main/src/mockit/internal/expectations/PhasedExecutionState.java
@@ -5,7 +5,6 @@
 package mockit.internal.expectations;
 
 import java.util.*;
-import java.util.Map.*;
 import javax.annotation.*;
 
 import mockit.internal.expectations.invocation.*;
@@ -16,16 +15,14 @@ final class PhasedExecutionState
 {
    @Nonnull final List<Expectation> expectations;
    @Nonnull final List<VerifiedExpectation> verifiedExpectations;
-   @Nonnull final Map<Object, Object> instanceMap;
-   @Nonnull final Map<Object, Object> replacementMap;
+   @Nonnull final EquivalentInstances equivalentInstances;
    @Nonnull final InstanceBasedMatching instanceBasedMatching;
    @Nullable PartiallyMockedInstances partiallyMockedInstances;
 
    PhasedExecutionState() {
       expectations = new ArrayList<>();
       verifiedExpectations = new ArrayList<>();
-      instanceMap = new IdentityHashMap<>();
-      replacementMap = new IdentityHashMap<>();
+      equivalentInstances = new EquivalentInstances();
       instanceBasedMatching = new InstanceBasedMatching();
    }
 
@@ -125,7 +122,7 @@ final class PhasedExecutionState
       InvocationArguments previousArguments = previousInvocation.arguments;
 
       if (newArguments.getMatchers() == null) {
-         return previousArguments.isMatch(newArguments.getValues(), instanceMap);
+         return previousArguments.isMatch(newArguments.getValues(), equivalentInstances.instanceMap);
       }
 
       return newArguments.hasEquivalentMatchers(previousArguments);
@@ -148,7 +145,7 @@ final class PhasedExecutionState
 
          if (
             isMatchingInvocation(mock, null, mockClassDesc, mockNameAndDesc, isConstructor, expectation) &&
-            expectation.invocation.arguments.isMatch(args, instanceMap)
+            expectation.invocation.arguments.isMatch(args, equivalentInstances.instanceMap)
          ) {
             if (expectation.recordPhase == null) {
                replayExpectationFound = expectation;
@@ -156,7 +153,7 @@ final class PhasedExecutionState
             }
 
             if (isConstructor) {
-               registerReplacementInstanceIfApplicable(mock, expectation.invocation);
+               equivalentInstances.registerReplacementInstanceIfApplicable(mock, expectation.invocation);
             }
 
             return expectation;
@@ -166,110 +163,25 @@ final class PhasedExecutionState
       return replayExpectationFound;
    }
 
-   private void registerReplacementInstanceIfApplicable(@Nullable Object mock, @Nonnull ExpectedInvocation invocation) {
-      Object replacementInstance = invocation.replacementInstance;
-
-      if (replacementInstance != null && replacementInstance != invocation.instance) {
-         replacementMap.put(mock, replacementInstance);
-      }
-   }
-
    private boolean isMatchingInstance(@Nonnull Object invokedInstance, @Nullable Boolean matchInstance, @Nonnull Expectation expectation) {
       ExpectedInvocation invocation = expectation.invocation;
       Object invocationInstance = invocation.instance;
       assert invocationInstance != null;
 
-      if (isEquivalentInstance(invocationInstance, invokedInstance)) {
+      if (equivalentInstances.isEquivalentInstance(invocationInstance, invokedInstance)) {
          return true;
       }
 
       if (
          TestRun.getExecutingTest().isInjectableMock(invokedInstance) ||
          partiallyMockedInstances != null && partiallyMockedInstances.isDynamicMockInstanceOrClass(invokedInstance, invocationInstance) ||
-         areNonEquivalentInstances(invocationInstance, invokedInstance)
+         equivalentInstances.areNonEquivalentInstances(invocationInstance, invokedInstance)
       ) {
          return false;
       }
 
       return
          (matchInstance == null || !matchInstance) && !invocation.matchInstance && expectation.recordPhase != null &&
-         !replacementMap.containsValue(invocationInstance);
-   }
-
-   boolean isEquivalentInstance(@Nonnull Object invocationInstance, @Nonnull Object invokedInstance) {
-      return
-         invocationInstance == invokedInstance ||
-         invocationInstance == replacementMap.get(invokedInstance) ||
-         invocationInstance == instanceMap.get(invokedInstance) ||
-         invokedInstance == instanceMap.get(invocationInstance);
-   }
-
-   private boolean areNonEquivalentInstances(@Nonnull Object invocationInstance, @Nonnull Object invokedInstance) {
-      boolean recordedInstanceMatchingAnyInstance = !isMatchingInstance(invocationInstance);
-      boolean invokedInstanceMatchingSpecificInstance = isMatchingInstance(invokedInstance);
-      return recordedInstanceMatchingAnyInstance && invokedInstanceMatchingSpecificInstance;
-   }
-
-   private boolean isMatchingInstance(@Nonnull Object instance) {
-      return
-         instanceMap.containsKey(instance)    || instanceMap.containsValue(instance) ||
-         replacementMap.containsKey(instance) || replacementMap.containsValue(instance);
-   }
-
-   boolean areInDifferentEquivalenceSets(@Nonnull Object mock1, @Nonnull Object mock2) {
-      if (mock1 == mock2 || instanceMap.isEmpty()) {
-         return false;
-      }
-
-      Object mock1Equivalent = instanceMap.get(mock1);
-      Object mock2Equivalent = instanceMap.get(mock2);
-
-      if (mock1Equivalent == mock2 || mock2Equivalent == mock1) {
-         return false;
-      }
-
-      //noinspection SimplifiableIfStatement
-      if (mock1Equivalent != null && mock2Equivalent != null) {
-         return true;
-      }
-
-      return instanceMapHasMocksInSeparateEntries(mock1, mock2);
-   }
-
-   private boolean instanceMapHasMocksInSeparateEntries(@Nonnull Object mock1, @Nonnull Object mock2) {
-      boolean found1 = false;
-      boolean found2 = false;
-
-      for (Entry<Object, Object> entry : instanceMap.entrySet()) {
-         if (!found1 && isInMapEntry(entry, mock1)) {
-            found1 = true;
-         }
-
-         if (!found2 && isInMapEntry(entry, mock2)) {
-            found2 = true;
-         }
-
-         if (found1 && found2) {
-            return true;
-         }
-      }
-
-      return false;
-   }
-
-   private static boolean isInMapEntry(@Nonnull Entry<Object, Object> mapEntry, @Nonnull Object mock) {
-      return mapEntry.getKey() == mock || mapEntry.getValue() == mock;
-   }
-
-   @Nullable
-   Object getReplacementInstanceForMethodInvocation(@Nonnull Object invokedInstance, @Nonnull String methodNameAndDesc) {
-      return methodNameAndDesc.charAt(0) == '<' ? null : replacementMap.get(invokedInstance);
-   }
-
-   boolean isReplacementInstance(@Nonnull Object invokedInstance, @Nonnull String methodNameAndDesc) {
-      return
-         methodNameAndDesc.charAt(0) != '<' && (
-            replacementMap.containsKey(invokedInstance) || replacementMap.containsValue(invokedInstance)
-         );
+         !equivalentInstances.replacementMap.containsValue(invocationInstance);
    }
 }
